Documentation for "Generate Test Cases from Figma Designs" (#604)
* Updated Screenshots in the "Projects and Settings" section.

Updated all the screenshots and GIFs in the "Projects and Settings" section according to the new UI.

* Updated screenshots in the Collaboration section and revamped Requirements page

Updated screenshots in the Collaboration section and revamped Requirements page according to the new UI.

* Dynamic Test Suites

Prepared a documentation for Dynamic Test Suites.

* Update dynamic-test-suites.md

* Documentation for "Generate Test Cases from Figma Designs"

I have created a documentation for "Generate Test Cases from Figma Designs".